CP-24324: add goheader and nolintlint linters, fix issues
The goheader linter is configured to identify files with missing
copyright headers. It also catches many instances where the the
copyright header is treated as part of the documentation. All the
issues it found have also been fixed.

nolintlint ensures that //nolint directives don't have a leading
space ("// nolint"; machine-readable comments aren't supposed to have
a space), include the specific linter(s) we want to disable, and don't
disable a linter unnecessarily.