CP-24405: add development tool installation support to Makefile
The obvious advantage of this is that people don't have to manually
install tools.

IMHO a more important advantage is that all the tools use a specific
version. The output of these tools can vary from version to version,
so making sure we keep the local versions synced with CI, and with
one another, will help us avoid unnecessarily conflicts.