CP-24407: Use `go generate` to (re)generate mocks, check in CI
I noticed that `make clean-mocks && make generate-mocks` wasn't
working, and after some digging I believe that is because of how the
mocks are generated; on a file by file basis instead of generating
specific interfaces.

Rather than rewrite the make rules to generate the mocks for the
interfaces, I've added //go:generate directives in the code, so
`go generate ./...` should do the right thing.

Getting the mocks to regenerate correctly requires a small patch to the
output for the ResourceStore, so instead of a note in the generated
code that there was a tweak I've added a diff which gets applied to the
output automatically.

Since we can now generate the mocks fully automatically, this means we
can add a check in CI to make sure that the generated mocks are up to
date, which I've also done in this patch.