CP-24413: add govulncheck to CI
govulncheck is an official tool used too check for known
vulnerabilities in dependencies. while it does duplicate some of the
functionality from dependabot, govulncheck actually analyzes the code
instead of just looking at the go.mod, so there should be very few
false positives, making it a good choice for something to run on every
PR, forcing us to take care of higher-priority updates even if
dependabot PRs are languishing.