CP-24425: replace /scrape HTTP endpoint with a -scrape command line option
For larger clusters, the /scrape endpoint was failing because it was
happening in the context of an HTTP handler, so eventually the context
was being cancelled.

The plan here is to replace the init job in the cloudzero-agent chart,
which previously called curl on the /scrape endpoint, to instead just
invoke the controller binary with the -scrape flag directly.