CP-25685: configure logger per settings, use it in the collector
This uses the logging configuration to create a Zerolog logger, and
sets it as the default context logger.

This logger is then picked up in the remote write API implementation
by middleware which is used to inject the logger, with some fields to
provide information about the request, into the request so that it can
be used in route handlers.