Rename shipper.File to shipper.MetricFile, and several symbols.
`MetricFile` makes it clear that the type should only be used to load
metrics files, which is important since the code now automatically
transcodes the file to Parquet as it is reading.

I've also renamed a few functions to better conform to go standards,
and updated several documentation strings to match godoc's
expectations, but there should be no functional changes in this
commit.