CP-25384: split metrics into cost and observability
This allows a user to specify filters to sort metrics into cost and
observability buckets, discard unwanted metrics, and discard unwanted
labels.

The filters can be passed in via the configuration files, and are
intended to be set in the Helm chart.

There are four sets of filters, cost name, cost labels, observability
name, and observability labels. The name filters perform matching on
the metric name and determine whether a metric will be sorted into
cost, observability, both, or be dropped. If they are not dropped, the
labels filter for the appropriate destination is executed to determine
whether each label should be retained or dropped.

Each filter can use a different matching strategy: exact, prefix,
suffix, contains, or regex. Exact matches are the preferred mechanism
as they are implemented using a hash table whereas all of the others
have to execute code to test each input against each filter.

The patch is large because many places in the code assumed we would
have a single store, but with this change we now have two stores.