CP-26896: reset ticker after flushing so we don't immediately flush
With the previous code, we just naively flushed every so often, even
if we had just flushed the metrics.  This mostly just resets the
ticker after each flush (regardless of what triggered the flush).

However, in order to make this work I had to move the ticker from
the metrics collector to the disk store, since the collector didn't
have any visibility into flushes triggered because of the row count.
I could have added some code to signal the collector from the store,
but honestly this feels more correct.