Always test Helm Chart in CI.
AFAICT the reason we weren't doing this before is that we didn't want
to publish the chart for any tag but develop, or a version tag.
However, the building and publishing is done in a different workflow.
That leads me to believe that this was just a condition left over from
a time when publishing was done in this workflow.