CP-33421: add GitHub Actions linter (actionlint)
This handy linter which is written in Go, and is intended to pick up on
potential issues with the GitHub Actions setup.

A lot of the noise from this commit comes from the fact that, if the
shellcheck utility is available (as it is in CI), actionlint will run
the shell commands through shellcheck, and that finds a lot of issues.
Mostly insufficient quoting of variables.